Separating backend interface to avoid dependency cycle
